fix: 레포지토리의 에러 발생 역할을 서비스로 위임

+++ apps/server/src/domain/user/service.ts
@@ -1,15 +1,18 @@
-import { FastifyReply, FastifyRequest, RouteHandler } from "fastify";
 import { userRepository } from "./repository";
 import {
   BaseResponseBuilder,
   ErrorResponseBuilder,
   SaveUserInfoReq,
+  UserError,
 } from "@ddanjit/domain";
 
 export const userService = {
   async getMe(email: string) {
     try {
       const user = await userRepository.findByEmail(email);
+      if (!user) {
+        throw new Error(UserError.NOTFOUND);
+      }
       return BaseResponseBuilder(200, undefined, user);
     } catch (e) {
       throw ErrorResponseBuilder(e);
@@ -19,6 +22,9 @@ export const userService = {
   async checkUserInfo(email: string) {
     try {
       const user = await userRepository.findByEmail(email);
+      if (!user) {
+        throw new Error(UserError.NOTFOUND);
+      }
       if (!user.name || !user.birthYear) {
         return BaseResponseBuilder(200, "유저 정보를 등록하지 않았습니다.", {
           registered: false,
@@ -35,12 +41,12 @@ export const userService = {
 
   async saveUserInfo(email: string, userInfo: SaveUserInfoReq) {
     try {
+      const isExist = await userRepository.findByEmail(email);
+      if (!isExist) {
+        throw new Error(UserError.NOTFOUND);
+      }
       const updatedUser = await userRepository.save(email, userInfo);
-      return BaseResponseBuilder(
-        201,
-        "정보가 저장되었어요.",
-        updatedUser,
-      );
+      return BaseResponseBuilder(201, "정보가 저장되었어요.", updatedUser);
     } catch (e) {
       throw ErrorResponseBuilder(e);
     }
@@ -48,6 +54,10 @@ export const userService = {
 
   async deleteUser(email: string) {
     try {
+      const isExist = await userRepository.findByEmail(email);
+      if (!isExist) {
+        throw new Error(UserError.NOTFOUND);
+      }
       await userRepository.delete(email);
       return BaseResponseBuilder(200, "그동안 즐거웠어요. 또 만나요!");
     } catch (e) {
